Fixed classes inheriting MapObject receiving only integers as parameters, when it should be doubles

+++ Brimstone/src/main/java/fi/make/brimstone/game/Enemy.java
@@ -4,7 +4,7 @@ import fi.make.brimstone.helpers.Variables;
 
 public class Enemy extends MapObject {
 
-    public Enemy(int x, int y) {
+    public Enemy(double x, double y) {
         super(x, y, Variables.ENEMY_IMAGE);
     }
 }
